Update the PR with comment fixes
