+++ src/sidenav/sidenav-context.tsx
@@ -1,12 +1,6 @@
-import React, {
-    Dispatch,
-    SetStateAction,
-    createContext,
-    useMemo,
-    useState,
-} from "react";
+import React, { Dispatch, SetStateAction, createContext } from "react";
 
-interface SelectedItem {
+export interface SelectedItem {
     itemId: string | undefined;
     openDrawer?: boolean | undefined;
 }
@@ -18,10 +12,6 @@ interface SidenavContextProps {
     setDrawerContent: Dispatch<SetStateAction<React.ReactNode | undefined>>;
 }
 
-interface SidenavProviderProps {
-    children: React.ReactNode;
-}
-
 export const SidenavContext = createContext<SidenavContextProps>({
     selectedItem: undefined,
     drawerContent: undefined,
@@ -31,30 +21,3 @@ export const SidenavContext = createContext<SidenavContextProps>({
         prevState: SetStateAction<React.ReactNode | undefined>
     ) => prevState,
 });
-
-export function SidenavProvider({
-    children,
-}: SidenavProviderProps): JSX.Element {
-    const [selectedItem, setSelectedItem] = useState<SelectedItem | undefined>(
-        undefined
-    );
-    const [drawerContent, setDrawerContent] = useState<
-        React.ReactNode | undefined
-    >(undefined);
-
-    const value = useMemo(
-        () => ({
-            selectedItem,
-            drawerContent,
-            setSelectedItem,
-            setDrawerContent,
-        }),
-        [selectedItem, drawerContent, setSelectedItem, setDrawerContent]
-    );
-
-    return (
-        <SidenavContext.Provider value={value}>
-            {children}
-        </SidenavContext.Provider>
-    );
-}
