+++ src/sidenav/sidenav.tsx
@@ -1,23 +1,92 @@
-import { Container } from "./sidenav.styles";
+import { useMemo, useRef, useState } from "react";
+import {
+    DesktopContainer,
+    DesktopDrawer,
+    MobileContainer,
+    Wrapper,
+} from "./sidenav.styles";
 import { SidenavProps } from "./types";
 import { SidenavItem } from "./sidenav-item";
 import { SidenavGroup } from "./sidenav-group";
-import { SidenavProvider } from "./sidenav-context";
+import { SelectedItem, SidenavContext } from "./sidenav-context";
 import { SidenavDrawerItem } from "./sidenav-drawer-item";
 import { SidenavDrawerSubitem } from "./sidenav-drawer-subitem";
-import { SideNavWrapper } from "./sidenav-wrapper";
+import { useSpring } from "react-spring";
+import { useEventListener } from "../util/use-event-listener";
 
 const SidenavBase = ({
     fixed = true,
     children,
     ...otherProps
 }: SidenavProps) => {
+    // =============================================================================
+    // CONST, STATE, REF
+    // =============================================================================
+    // const { drawerContent, selectedItem, setDrawerContent, setSelectedItem } =
+    //     useContext(SidenavContext);
+    const wrapperRef = useRef<HTMLDivElement>(null);
+
+    const [selectedItem, setSelectedItem] = useState<SelectedItem | undefined>(
+        undefined
+    );
+    const [drawerContent, setDrawerContent] = useState<
+        React.ReactNode | undefined
+    >(undefined);
+
+    const value = useMemo(
+        () => ({
+            selectedItem,
+            drawerContent,
+            setSelectedItem,
+            setDrawerContent,
+        }),
+        [selectedItem, drawerContent, setSelectedItem, setDrawerContent]
+    );
+
+    const drawerAnimationProps = useSpring({
+        width: drawerContent ? 240 : 0,
+    });
+
+    // =========================================================================
+    // EVENT HANDLERS
+    // =========================================================================
+    const handleOutsideClicks = (e: PointerEvent) => {
+        if (
+            e.target instanceof HTMLDivElement &&
+            wrapperRef.current &&
+            !wrapperRef.current.contains(e.target)
+        ) {
+            setSelectedItem({ ...selectedItem, openDrawer: false });
+            setDrawerContent(undefined);
+        }
+    };
+
+    const handleMouseLeave = () => {
+        setDrawerContent(undefined);
+    };
+
+    useEventListener("click", handleOutsideClicks);
+
     return (
-        <Container $fixed={fixed} {...otherProps}>
-            <SidenavProvider>
-                <SideNavWrapper>{children}</SideNavWrapper>
-            </SidenavProvider>
-        </Container>
+        <SidenavContext.Provider value={value}>
+            <Wrapper
+                $fixed={fixed}
+                {...otherProps}
+                ref={wrapperRef}
+                onMouseLeave={handleMouseLeave}
+            >
+                <DesktopContainer {...otherProps}>{children}</DesktopContainer>
+                <DesktopDrawer
+                    style={drawerAnimationProps}
+                    $showDrawer={!!drawerContent}
+                    data-testid="sidenav-drawer"
+                >
+                    {drawerContent}
+                </DesktopDrawer>
+                {/** NOTE: Since mobile view not supported yet, children will not be rendered */}
+                <MobileContainer></MobileContainer>
+            </Wrapper>
+        </SidenavContext.Provider>
     );
 };
